+++ test-project/Packages/io.improbable.gdk.dependencytest/.codegen/Source/Generators/TestProject/ModularCodegenTestGenerator.cs
@@ -1,0 +1,20 @@
+using Improbable.Gdk.CodeGeneration.CodeWriter;
+
+namespace Improbable.Gdk.CodeGenerator
+{
+    public static class ModularCodegenTestGenerator
+    {
+        public static string Generate()
+        {
+            return CodeWriter.Populate(cgw =>
+            {
+                cgw.Namespace("Improbable.Gdk.ModularCodegenTests", ns =>
+                {
+                    ns.Type("public class TemplateTest", t =>
+                    {
+                    });
+                });
+            }).Format();
+        }
+    }
+}